% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allpairs_proportionbf}
\alias{allpairs_proportionbf}
\title{Automates Bayes Factor proportion test on all pairs of groups in a dataframe}
\usage{
allpairs_proportionbf(
  df,
  group_col = "NA",
  count_col = "NA",
  sample_type = "indepMulti"
)
}
\arguments{
\item{df}{dataframe}

\item{group_col}{name of column with group names}

\item{sample_type}{sampling scheme, one of `c("indepMulti", "jointMulti", "poisson")` default = "indepMulti"}

\item{data_col}{name of column with measurements in}
}
\value{
dataframe with 7 columns
\enumerate{
\item `control_group` - the name of the group considered control
\item. `test_group` - the group considered test
\item `h_0` - statement of the $H_0$ hypothesis
\item `h_1` - statement of the $H_1$ hypothesis
\item `BayesFactor` - the resulting Bayes Factor from \eqn{H_{1}/H_{0}}
\item `odds_h_1` - statement of the Bayes Factor in terms of odds $H_0:H_1$
\item `summary` - summary statement of relative evidence for the hypothesis
}
}
\description{
Performs a Bayes factor proportoin test for all pairs of groups in a named group column in the provided
dataframe. Group pairings are made in one direction only since order makes no difference.
A thin wrapper around the `BayesFactor::contingencyTableBF()` function. Always set up so that effect size is test - control.
}
\examples{
allpairs_ttestbf(PlantGrowth, group_col='group', data_col='weight')
}
