% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allpairs_ttestbf}
\alias{allpairs_ttestbf}
\title{Automates Bayes Factor \eqn{t}-test on all pairs of groups in a dataframe}
\usage{
allpairs_ttestbf(
  df,
  group_col = NA,
  data_col = NA,
  h_1 = "test_greater_than_control",
  rscale = "medium"
)
}
\arguments{
\item{df}{dataframe}

\item{group_col}{name of column with group names}

\item{data_col}{name of column with measurements in}

\item{h_1}{the h_1 hypothesis, one of `c("test_greater_than_control", "test_smaller_than_control","test_not_equal_to_control")` default = "test_greater_than_control"}

\item{rscale}{scale for the prior, one of `c("medium", "wide", "ultrawide")` default = "medium"}
}
\value{
dataframe with 7 columns
\enumerate{
\item `control_group` - the name of the group considered control
\item. `test_group` - the group considered test
\item `h_0` - statement of the $H_0$ hypothesis
\item `h_1` - statement of the $H_1$ hypothesis
\item `BayesFactor` - the resulting Bayes Factor from \eqn{H_{1}/H_{0}}
\item `odds_h_1` - statement of the Bayes Factor in terms of odds $H_0:H_1$
\item `summary` - summary statement of relative evidence for the hypothesis
}
}
\description{
Performs a Bayes factor $t$-test for all pairs of groups in a named group column in the provided
dataframe. Group pairings are made in both directions 'grp1 - grp2' and 'grp2 - grp1'.
A thin wrapper around the `BayesFactor::ttestBF()` function. Always set up so that effect size is test - control.
}
\examples{
allpairs_ttestbf(PlantGrowth, group_col='group', data_col='weight')
}
