% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{named_pair_proportionbf}
\alias{named_pair_proportionbf}
\title{Perform a Bayes Factor contingency table test on two named groups in a dataframe}
\usage{
named_pair_proportionbf(
  df,
  group_col = NA,
  count_col = NA,
  control = NA,
  test = NA,
  sample_type = "indepMulti"
)
}
\arguments{
\item{df}{dataframe}

\item{group_col}{name of column with group names}

\item{control}{name of control group}

\item{test}{name of test group}

\item{sample_type}{sampling scheme, one of `c("indepMulti", "jointMulti", "poisson")` default = "indepMulti"}

\item{data_col}{name of column with measurements in}
}
\value{
dataframe with one row and 7 columns
\enumerate{
\item `control_group` - the name of the group considered control
\item. `test_group` - the group considered test
\item `h_0` - statement of the $H_0$ hypothesis
\item `h_1` - statement of the $H_1$ hypothesis
\item `BayesFactor` - the resulting Bayes Factor from \eqn{H_{1}/H_{0}}
\item `odds_h_1` - statement of the Bayes Factor in terms of odds $H_0:H_1$
\item `summary` - summary statement of relative evidence for the hypothesis
}
}
\description{
Performs a Bayes factor contingency table test on a named test and control group in the provided
dataframe. A thin wrapper around the `BayesFactor::contingencyTableBF()` function. Always set up so that effect size is test - control.
}
\examples{
named_pair_proportionbf(PlantGrowth, group_col='group', data_col='weight', control='crtl', test='trt2')
}
