% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{named_pair_ttestbf}
\alias{named_pair_ttestbf}
\title{Perform a Bayes Factor \eqn{t}-test on two named groups in a dataframe}
\usage{
named_pair_ttestbf(
  df,
  group_col = NA,
  data_col = NA,
  control = NA,
  test = NA,
  h_1 = "test_greater_than_control",
  rscale = "medium"
)
}
\arguments{
\item{df}{dataframe}

\item{group_col}{name of column with group names}

\item{data_col}{name of column with measurements in}

\item{control}{name of control group}

\item{test}{name of test group}

\item{h_1}{the h_1 hypothesis, one of `c("test_greater_than_control", "test_smaller_than_control","test_not_equal_to_control")` default = "test_greater_than_control"}

\item{rscale}{scale for the prior, one of `c("medium", "wide", "ultrawide")` default = "medium"}
}
\value{
dataframe with one row and 7 columns
\enumerate{
\item `control_group` - the name of the group considered control
\item. `test_group` - the group considered test
\item `h_0` - statement of the $H_0$ hypothesis
\item `h_1` - statement of the $H_1$ hypothesis
\item `BayesFactor` - the resulting Bayes Factor from \eqn{H_{1}/H_{0}}
\item `odds_h_1` - statement of the Bayes Factor in terms of odds $H_0:H_1$
\item `summary` - summary statement of relative evidence for the hypothesis
}
}
\description{
Performs a Bayes factor $t$-test on a named test and control group in the provided
dataframe. A thin wrapper around the `BayesFactor::ttestBF()` function. Always set up so that effect size is test - control.
}
\examples{
named_pair_ttestbf(PlantGrowth, group_col='group', data_col='weight', control='crtl', test='trt2')
}
